Remove assessment process code since it will be part of other PR
