--- src/pages/application-inventory/application-assessment/application-assessment.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const ApplicationAssessment: React.FC = () => {
-  return <>Work in progres</>;
-};
