--- src/pages/application-inventory/application-assessment/index.ts
@@ -1,1 +1,0 @@
-export { ApplicationAssessment as default } from "./application-assessment";
