--- src/pages/application-inventory/application-list/index.ts
@@ -1,1 +1,0 @@
-export { ApplicationList as default } from "./application-list";
