+++ src/pages/application-inventory/components/application-form/application-form.tsx
