+++ src/pages/application-inventory/components/application-form/index.ts
