+++ src/pages/application-inventory/components/application-form/stories/application-form.stories.tsx
