+++ src/pages/application-inventory/components/new-application-modal/new-application-modal.tsx
