+++ src/pages/application-inventory/components/remote-assessment/index.ts
