+++ src/pages/application-inventory/components/remote-assessment/remote-assessment.tsx
