+++ src/pages/application-inventory/components/toolbar-search-filter/input-text-filter/input-text-filter.tsx
