+++ src/pages/application-inventory/components/update-application-modal/index.ts
